# PERF: Define PeriodArray._values_for_argsort